requirements.txt
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
celery-stubs==0.1.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.4
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
filelock==3.15.4
flake8==7.1.0
flower==2.0.1
frozenlist==1.4.1
gunicorn==22.0.0
humanize==4.10.0
identify==2.5.36
idna==3.7
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.7
mccabe==0.7.0
multidict==6.0.5
mypy==1.10.1
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg==3.1.19
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tornado==6.4.1
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240622
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
yarl==1.9.4
